moved pushNotification from polling to it's own function.
